Move brand helper to Scope
